pyproject.toml
[tool.poetry]
name = "mqtttest"
version = "0.1.0"
description = ""
authors = ["Ahmed Shariff <shariff.mfa@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
